Bug fix in the item-item recsys and added eval code
The item-item recsys works well when you have a threshold on the item
similarity. That is, if two items have a similarity of 0.1, should the
rating of one item really contribute to that other one? Added a
threshold to prevent this.

New scaffold code as well for handling hold-out data